requirements.txt
biopython==1.72
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
pandas==0.23.4
requests==2.21.0
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
statsmodels==0.9.0
urllib3==1.24.1
XlsxWriter==1.1.2